add new endpoint to API
