+++ website/src/components/SearchPage/SearchForm.tsx
@@ -195,7 +195,6 @@ const SearchField = ({ field, lapisUrl, fieldValues, setSomeFieldValues, lapisSe
                         fieldValue={(fieldValues[field.name] ?? '') as string}
                         setSomeFieldValues={setSomeFieldValues}
                         lapisUrl={lapisUrl}
-                        lapisSearchParameters={lapisSearchParameters}
                     />
                 );
             }
